+++ common/src/plugins/resource/archive/abstract_archive_stream_reader.cpp
@@ -146,7 +146,7 @@ void QnAbstractArchiveReader::run()
             data->dataProvider = this;
 
         if (videoData)
-            m_stat[videoData->channelNumber].onData(videoData->dataSize());
+            m_stat[videoData->channelNumber].onData(static_cast<unsigned int>(videoData->dataSize()));
 
 
         putData(std::move(data));
